Move `duration_secs` from MIDI config to window
Quite often, I'll change the on-screen duration depending on if it's
portrait or landscape, so another small change I'll benefit from a lot.

It defaults to 5 seconds so for WAV renders you can leave it out of your
window config and it'll be ignored, or you can use the default for MIDI
as well!