Add lyrics to the bottom of video
If you provide `lyrics_file` to your midi.json, it'll load the LRC file
and parse the lyrics to be used on-screen. Do keep in mind that it's
currently programmed to assume each line is in chronological order, and
doesn't support multiple timestamps for repeated lines.

I would really like to support coloured text, but LRC files are by far
the simplest format to work with, and I absolutely do not fancy parsing
XML or other, more complicated formats (for now)!